requirements.txt
matplotlib==3.10.1
numpy==2.1.3
pandas==2.2.3
pyinstrument==5.0.1
pyroomacoustics==0.8.3
pystoi==0.4.1
PyYAML==6.0.2
resampy==0.4.3
scipy==1.15.2
silero-vad==5.1.2
soundfile==0.13.1
tqdm==4.67.1
